Re-architecture project
Fix #138
