+++ src/command-handlers/index.ts
@@ -1,2 +1,10 @@
+export * from './configure';
 export * from './devices';
+export * from './logout';
+export * from './passwords';
+export * from './secureNotes';
+export * from './sync';
 export * from './teamDevices';
+export * from './teamLogs';
+export * from './teamMembers';
+export * from './teamReport';
