+++ src/endpoints/index.ts
@@ -1,14 +1,17 @@
 export * from './completeDeviceRegistration';
+export * from './completeLoginWithAuthTicket';
+export * from './deactivateDevices';
+export * from './deactivateTeamDevice';
+export * from './getAuditLogs';
 export * from './getLatestContent';
 export * from './getStatus';
 export * from './getTeamMembers';
+export * from './getTeamReport';
+export * from './listDevices';
+export * from './listTeamDevices';
 export * from './performDashlaneAuthenticatorVerification';
 export * from './performDuoPushVerification';
 export * from './performEmailTokenVerification';
 export * from './performTotpVerification';
 export * from './requestDeviceRegistration';
-export * from './getAuditLogs';
-export * from './getTeamReport';
-export * from './listDevices';
-export * from './listTeamDevices';
-export * from './deactivateDevices';
+export * from './registerTeamDevice';
