--- src/middleware/index.ts
@@ -1,9 +1,0 @@
-export * from './configure';
-export * from './getAuditLogs';
-export * from './getPasswords';
-export * from './getSecureNotes';
-export * from './getTeamMembers';
-export * from './getTeamReport';
-export * from './registerDevice';
-export * from './reset';
-export * from './sync';
