+++ src/modules/auth/index.ts
@@ -1,0 +1,2 @@
+export * from './perform2FAVerification';
+export * from './registerDevice';
