+++ src/modules/crypto/encrypt.ts
