+++ src/modules/crypto/hash.ts
