+++ src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './dialogs';
 export * from './gotImplementation';
 export * from './strings';
+export * from './teamDeviceCredentials';
